requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==21.2.0
    # via
    #   jsonschema
    #   markdown-it-py
click==8.0.1
    # via -r requirements.in
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   traitlets
jsonschema==3.2.0
    # via nbformat
jupyter-core==4.7.1
    # via nbformat
jupytext==1.11.4
    # via -r requirements.in
markdown-it-py==1.1.0
    # via
    #   jupytext
    #   mdit-py-plugins
mdit-py-plugins==0.2.8
    # via jupytext
nbformat==5.1.3
    # via jupytext
pyrsistent==0.18.0
    # via jsonschema
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   jupytext
six==1.16.0
    # via jsonschema
toml==0.10.2
    # via jupytext
traitlets==5.0.5
    # via
    #   jupyter-core
    #   nbformat

# The following packages are considered to be unsafe in a requirements file:
# setuptools
